Use rvalue reference for CFArray arg to set_convolution_function()
Force caller to relinquish control of value. Not strictly necessary
here as there's always intermediate copy on the host, but this is a
pattern I intend to follow elsewhere. Putting the rvalue argument at
the top level ensures that a minimal number of copies is made, i.e,
it's the caller's responsibility.