Fix for runtime error with one data polarization
Fixed error could be invoked on the experimental
VisibilityGridder::kernel() using minigridder with "--kernel 1
--mueller I1 --repeats 2 --op both". I think that the repeat parameter
forced a reuse of CUDA streams, which was a necessary condition. Also,
any Mueller matrix for N=1 also triggered the failure, e.g
{{0},{1}}. The "both" op parameter value was necessary, as there was
no error for any other op value. The fix is implemented by ensuring
that the "visbuff" View is captured by the lambda in the parallel_for;
hints for the fix were provided by cuda-gdb and Kokkos error output,
but very indirectly.