Change value element type in impl::WritableDeviceCFArray
This change ensures that the alignment differences in Kokkos::Complex and
std::complex are accounted for properly, at the cost of needing to make copies
of value vectors that are provided in a constructor argument. In the future, it
may be worth investigating how it might be possible to define the aligned types
in the hpg namespace (and top-level API) without exposing the Kokkos types in
the implementation.